Rename unit tests to match convention; add UI tests; improve error message when type is not found in referenced assemblies
