+++ Xamarin.Forms.Controls.Issues/Xamarin.Forms.Controls.Issues.CustomNamespaceUris/Issue2691/CustomButton.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+using Xamarin.Forms;
+
+namespace Xamarin.Forms.Controls.CustomNamespace1
+{	
+    public class CustomButton : Button
+    {
+    }
+}
