+++ Xamarin.Forms.Controls.Issues/Xamarin.Forms.Controls.Issues.CustomNamespaceUris/Issue2691/CustomLabel.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+using Xamarin.Forms;
+
+namespace Xamarin.Forms.Controls.CustomNamespace2
+{
+    public class CustomLabel : Label
+    {
+    }
+}
