+++ Xamarin.Forms.Controls.Issues/Xamarin.Forms.Controls.Issues.CustomNamespaceUris/Issue2691/Library.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Xamarin.Forms.Controls.Issues.CustomNamespaceUris
+{
+    public class Library
+    {
+		public static void Init()
+		{
+
+		}
+    }
+}
