+++ Xamarin.Forms.Controls.Issues/Xamarin.Forms.Controls.Issues.Shared/Github2691.xaml
@@ -1,0 +1,17 @@
+﻿<?xml version="1.0" encoding="UTF-8"?>
+<ContentPage xmlns="http://xamarin.com/schemas/2014/forms" 
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+			 xmlns:cc1="http://xamarin.com/schemas/2014/forms/customurl1"
+			 xmlns:cc2="http://xamarin.com/schemas/2014/forms/customurl2"
+             x:Class="Xamarin.Forms.Controls.Issues.Github2691">
+	<ContentPage.Content>
+		<StackLayout>
+			<cc1:CustomButton Text="Custom Button"/>
+			<cc1:CustomLabel Text="Custom Label"/>
+			<cc2:CustomStackLayout>
+				<Label Text="Label inside Custom StackLayout"/>
+				<Label Text="Another Label inside Custom StackLayout"/>
+			</cc2:CustomStackLayout>
+		</StackLayout>
+	</ContentPage.Content>
+</ContentPage>
