+++ Xamarin.Forms.Controls/Xamarin.Forms.Controls.csproj
@@ -26,6 +26,7 @@
   </ItemGroup>
   <ItemGroup>
     <!-- A reference to the entire .NET Framework is automatically included -->
+    <ProjectReference Include="..\Xamarin.Forms.Controls.Issues\Xamarin.Forms.Controls.Issues.CustomNamespaceUris\Xamarin.Forms.Controls.Issues.CustomNamespaceUris.csproj" />
     <ProjectReference Include="..\Xamarin.Forms.Core\Xamarin.Forms.Core.csproj" />
     <ProjectReference Include="..\Xamarin.Forms.CustomAttributes\Xamarin.Forms.CustomAttributes.csproj" />
     <ProjectReference Include="..\Xamarin.Forms.Maps\Xamarin.Forms.Maps.csproj" />
